Factor the Package, Override and Diff Report code
This pulls the code for generating a package report, both styles of
override reports and the bulk diff of packages out into their own
command files.

Each command is placed in its own self contained file with the command
and the thread code responsible for generating it.
