+++ src/__init__.py
@@ -19,11 +19,7 @@ __all__ = [
     "OverrideAuditEditSettingsCommand",
 
     # override_audit
-    "OverrideAuditPackageReportCommand",
-    "OverrideAuditOverrideReportCommand",
-    "OverrideAuditDiffPackageCommand",
     "OverrideAuditDiffOverrideCommand",
-    "OverrideAuditModifyMarkCommand",
     "OverrideAuditContextOverrideCommand",
     "OverrideAuditContextPackageCommand",
     "OverrideAuditContextReportCommand",
@@ -33,5 +29,8 @@ __all__ = [
     "OverrideAuditContextListener",
 
     # commands/*
+    "OverrideAuditPackageReportCommand",
+    "OverrideAuditOverrideReportCommand",
+    "OverrideAuditDiffPackageCommand",
     "OverrideAuditModifyMarkCommand"
 ]
