+++ src/commands/__init__.py
@@ -1,9 +1,19 @@
 from ...bootstrap import reload
 
-reload("src.commands", ["modify_mark"])
+reload("src.commands", ["package_report", "override_report", "diff_package",
+       "modify_mark"])
 
+from .package_report import OverrideAuditPackageReportCommand
+from .override_report import OverrideAuditOverrideReportCommand
+from .diff_package import OverrideAuditDiffPackageCommand
 from .modify_mark import OverrideAuditModifyMarkCommand
 
 __all__ = [
+    # Reports
+    "OverrideAuditPackageReportCommand",
+    "OverrideAuditOverrideReportCommand",
+    "OverrideAuditDiffPackageCommand",
+
+    # General
     "OverrideAuditModifyMarkCommand"
 ]
