+++ src/override_audit.py
@@ -566,285 +566,6 @@ class ReportGenerationThread(BackgroundWorkerThread):
 ###----------------------------------------------------------------------------
 
 
-class PackageReportThread(ReportGenerationThread):
-    """
-    Generate a tabular report of all installed packages and their state.
-    """
-    def _process(self):
-        pkg_list = PackageList()
-        pkg_counts = pkg_list.package_counts()
-
-        title = "{} Total Packages".format(len(pkg_list))
-        t_sep = "=" * len(title)
-
-        fmt = '{{:>{}}}'.format(len(str(max(pkg_counts))))
-        stats = ("{0} [S]hipped with Sublime\n"
-                 "{0} [I]nstalled (user) sublime-package files\n"
-                 "{0} [U]npacked in Packages\\ directory\n"
-                 "{0} Currently in ignored_packages\n"
-                 "{0} Installed Dependencies\n").format(fmt).format(*pkg_counts)
-
-        row = "| {:<40} | {:3} | {:3} | {:<3} |".format("", "", "", "")
-        r_sep = "-" * len(row)
-
-        result = [title, t_sep, "", self._generation_time(), stats, r_sep]
-        for pkg_name, pkg_info in pkg_list:
-            result.append(
-                "| {:<40} | [{:1}] | [{:1}] | [{:1}] |".format(
-                    decorate_pkg_name(pkg_info, name_only=True),
-                    "S" if pkg_info.shipped_path is not None else " ",
-                    "I" if pkg_info.installed_path is not None else " ",
-                    "U" if pkg_info.unpacked_path is not None else " "))
-        result.extend([r_sep, ""])
-
-        self._set_content("OverrideAudit: Package Report", result, ":packages",
-                          oa_syntax("OA-PkgReport"))
-
-
-###----------------------------------------------------------------------------
-
-
-class OverrideReportThread(ReportGenerationThread):
-    """
-    Generate a report on all packages which have overrides and what they are,
-    if any. The report always includes expired packages and overrides, but the
-    optional parameter filters to only show expired results.
-    """
-    def _process(self):
-        pkg_list = PackageList()
-
-        ignored = oa_setting("ignore_overrides_in")
-
-        only_expired = self.args["only_expired"]
-        ignore_empty = self.args["ignore_empty"]
-
-        if only_expired:
-            title = "OverrideAudit: Expired Override Report"
-            report_type = ":overrides_expired"
-        else:
-            title = "OverrideAudit: Override Report"
-            report_type = ":overrides"
-
-        expired_pkgs = []
-        result = []
-        if only_expired:
-            result.append("WARNING: Showing only expired overrides!\n"
-                          "WARNING: Non-expired overrides may exist!\n")
-        result.append(self._generation_time())
-
-        displayed = 0
-        for pkg_name, pkg_info in pkg_list:
-            if pkg_name not in ignored:
-                if self._output_package(result, pkg_info, only_expired,
-                                        expired_pkgs):
-                    displayed += 1
-
-        if displayed == 0:
-            if ignore_empty:
-                return sublime.set_timeout(self._notify_empty(), 10)
-
-            result.append(self._empty_msg())
-
-        self._set_content(title, result, report_type,
-                          oa_syntax("OA-OverrideReport"),
-                          {"override_audit_expired_pkgs": expired_pkgs})
-
-    def _output_package(self, result, pkg_info, only_expired, expired_pkgs):
-        shipped_override = pkg_info.has_possible_overrides(simple=False)
-        normal_overrides = pkg_info.override_files(simple=True)
-
-        expired_overrides = pkg_info.expired_override_files(simple=True)
-        expired_pkg = bool(pkg_info.expired_override_files(simple=False))
-
-        # No need to do anything if there are no overrides at all
-        if not normal_overrides and not shipped_override:
-            return False
-
-        if only_expired and not expired_overrides and not expired_pkg:
-            return False
-
-        if expired_overrides:
-            expired_pkgs.append(pkg_info.name)
-
-        result.append(decorate_pkg_name(pkg_info))
-
-        self._output_overrides(result, normal_overrides,
-                               expired_overrides, only_expired)
-        result.append("")
-
-        return True
-
-    def _output_overrides(self, result, overrides, expired, only_expired):
-        if not overrides:
-            return result.append("    [No simple overrides found]")
-
-        # Must be overrides, if none are expired use a different message.
-        if only_expired and not expired:
-            return result.append("    [No expired simple overrides found]")
-
-        for item in (expired if only_expired else overrides):
-            fmt = "  `- {}" if item not in expired else "  `- [X] {}"
-            result.append(fmt.format(item))
-
-    def _empty_msg(self):
-        return "No packages with %soverrides found" % (
-                "expired " if self.args["only_expired"] else "")
-
-    def _notify_empty(self):
-        log(self._empty_msg(), status=True)
-
-
-###----------------------------------------------------------------------------
-
-
-class BulkDiffReportThread(ReportGenerationThread):
-    """
-    Perform a bulk diff of all overrides in either all packages or a single
-    package, depending on the argument provided.
-
-    This is invoked from OverrideAuditDiffOverride when you invoke that command
-    with the bulk argument set to true.
-    """
-    def _process(self):
-        package = self.args["package"]
-        force_reuse = self.args["force_reuse"]
-
-        pkg_list = PackageList(package)
-
-        if package is not None:
-            if package not in pkg_list:
-                return log("Cannot diff package '%s'; not found" % package,
-                            status=True, dialog=True)
-
-            items = [package]
-        else:
-            items = packages_with_overrides(pkg_list)
-
-        self._diff_packages(items, pkg_list, package is not None, force_reuse)
-
-    def _diff_packages(self, names, pkg_list, single_package, force_reuse):
-        context_lines = oa_setting("diff_context_lines")
-        binary_patterns = oa_setting("binary_file_patterns")
-
-        result = []
-        title = "Override Diff Report: "
-        description = "Bulk Diff Report for overrides in"
-        report_type = ":bulk_all"
-        expired_pkgs = []
-
-        if len(names) == 1 and single_package:
-            title += names[0]
-            result.append(description + " {}\n".format(names[0]))
-            report_type = names[0]
-        elif len(names) == 0:
-            title += "All Packages"
-            result.append("No packages with overrides found to diff\n")
-        else:
-            title += "All Packages"
-            result.append(description + " {} packages\n".format(len(names)))
-
-        result.append(self._generation_time())
-
-        for name in names:
-            pkg_info = pkg_list[name]
-            if binary_patterns is not None:
-                pkg_info.set_binary_pattern(binary_patterns)
-
-            result.append(decorate_pkg_name(pkg_info))
-            self._perform_diff(pkg_info, context_lines, result,
-                               expired_pkgs)
-
-        self._set_content(title, result, report_type, oa_syntax("OA-Diff"),
-                          {"override_audit_expired_pkgs": expired_pkgs})
-
-    def _perform_diff(self, pkg_info, context_lines, result, expired_pkgs):
-        override_list = pkg_info.override_files(simple=True)
-        expired_list = pkg_info.expired_override_files(simple=True)
-        empty_diff_hdr = oa_setting("diff_empty_hdr")
-
-        if expired_list:
-            expired_pkgs.append(pkg_info.name)
-
-        for file in override_list:
-            if file in expired_list:
-                result.append("    [X] {}".format(file))
-            else:
-                result.append("    {}".format(file))
-
-            diff = pkg_info.override_diff(file, context_lines,
-                                          empty_result="No differences found",
-                                          binary_result="<File is binary>",
-                                          indent=8)
-
-            if diff is None:
-                content = (" " * 8) + ("Error opening or decoding file;"
-                                       " is it UTF-8 or Binary?")
-            else:
-                prefix = diff.hdr if diff.is_empty and empty_diff_hdr else ""
-                content = prefix + diff.result
-
-            result.extend([content, ""])
-
-        if len(override_list) == 0:
-            if pkg_info.has_possible_overrides(simple=True):
-                result.append("    [No simple overrides found]")
-            else:
-                reason = ("no sublime-package" if pkg_info.is_unpacked() else
-                          "no unpacked files")
-                result.append("    [No overrides possible; %s]" % reason)
-        result.append("")
-
-
-###----------------------------------------------------------------------------
-
-
-class OverrideAuditPackageReportCommand(sublime_plugin.WindowCommand):
-    """
-    Generate a tabular report of all installed packages and their state.
-    """
-    def run(self, force_reuse=False):
-        PackageReportThread(self.window, "Generating Package Report",
-                            self.window.active_view(),
-                            force_reuse=force_reuse).start()
-
-
-###----------------------------------------------------------------------------
-
-
-class OverrideAuditOverrideReportCommand(sublime_plugin.WindowCommand):
-    """
-    Generate a report on all packages which have overrides and what they are,
-    if any. The report always includes expired packages and overrides, but the
-    optional parameter filters to only show expired results.
-    """
-    def run(self, force_reuse=False, only_expired=False, ignore_empty=False):
-        OverrideReportThread(self.window, "Generating Override Report",
-                             self.window.active_view(),
-                             force_reuse=force_reuse,
-                             only_expired=only_expired,
-                             ignore_empty=ignore_empty).start()
-
-
-###----------------------------------------------------------------------------
-
-
-class OverrideAuditDiffPackageCommand(sublime_plugin.WindowCommand):
-    """
-    Perform a bulk diff of all overrides in either all packages or a single
-    package, depending on the argument provided.
-
-    This is invoked from OverrideAuditDiffOverride when you invoke that command
-    with the bulk argument set to true.
-    """
-    def run(self, package=None, force_reuse=False):
-        BulkDiffReportThread(self.window, "Generating Bulk Diff",
-                             self.window.active_view(),
-                             package=package, force_reuse=force_reuse).start()
-
-
-###----------------------------------------------------------------------------
-
-
 class OverrideAuditDiffOverrideCommand(sublime_plugin.WindowCommand):
     """
     Selective diff of a single override or package based on the provided
